+++ packages/vision_ai/lib/src/models/emotion.dart
@@ -1,3 +1,4 @@
+/// Detected facial emotion types (FER2013 universal emotions).
 enum Emotion {
   happy,
   sad,
@@ -8,5 +9,6 @@ enum Emotion {
   neutral,
   none;
 
+  /// Whether an emotion was detected (not [none]).
   bool get isRecognized => this != none;
 }